Refactor the rest of the old methods
CustomBot now handles all basic additional behavior, with the exception
of external "pretty" methods used for display. These need to be
addressed in the future.

Permissions are not implemented; this build should not be used in a real
classroom setting as any users with the commands could set the
attendance secret and see the attendance list.